+++ programs/MushroomClassifier.php
@@ -13,11 +13,12 @@ use Rubix\ML\Transformers\OneHotEncoder;
 use Rubix\ML\Reports\MulticlassBreakdown;
 use Rubix\ML\Transformers\LambdaFunction;
 use Rubix\ML\CrossValidation\Metrics\MCC;
+use Rubix\ML\NeuralNet\Layers\AlphaDropout;
 use Rubix\ML\Transformers\ZScaleStandardizer;
 use Rubix\ML\Classifiers\MultiLayerPerceptron;
 use Rubix\ML\Transformers\SparseRandomProjector;
-use Rubix\ML\NeuralNet\CostFunctions\RelativeEntropy;
-use Rubix\ML\NeuralNet\ActivationFunctions\ThresholdedReLU;
+use Rubix\ML\NeuralNet\ActivationFunctions\SELU;
+use Rubix\ML\NeuralNet\CostFunctions\CrossEntropy;
 use League\Csv\Reader;
 
 echo '╔═════════════════════════════════════════════════════╗' . "\n";
@@ -44,18 +45,16 @@ $labels = iterator_to_array($reader->fetchColumn('class'));
 
 $dataset = new Labeled($samples, $labels);
 
-$hidden = [
-    new Dense(20, new ThresholdedReLU(0.0)),
-    new Dense(25, new ThresholdedReLU(0.0)),
-    new Dense(20, new ThresholdedReLU(0.0)),
-];
-
-$estimator = new Pipeline(new MultiLayerPerceptron($hidden, 50, new Adam(0.001),
-    1e-4, new RelativeEntropy(), 1e-3, new MCC(), 0.1, 3, 100), [
-        new OneHotEncoder(),
-        new SparseRandomProjector(30),
-        new ZScaleStandardizer(),
-    ]);
+$estimator = new Pipeline(new MultiLayerPerceptron([
+    new AlphaDropout(40, 0.1),
+    new Dense(30, new SELU()),
+    new AlphaDropout(20, 0.1),
+    new Dense(10, new SELU()),
+], 100, new Adam(0.001), 1e-4, new CrossEntropy(), 1e-3, new MCC(), 0.1, 3, 100), [
+    new OneHotEncoder(),
+    new SparseRandomProjector(30),
+    new ZScaleStandardizer(),
+]);
 
 $report = new AggregateReport([
     new ConfusionMatrix(),